+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## Current
+## 0.0.14
 - :rocket: updated wdio version to 8
 - :rocket: added I press Y key X times step
 - :rocket: added I execute steps allow to execute client functions
+- :grey_question: added workaround to use keys in wdio8
 
 ## 0.0.13
 - :rocket: added I wait until current url step